base_url is just the planet base url, use f-strings for url building, remove ref to auth in example
The orders API includes multiple api endpoints. They all begin with
'https://api.planet.com/' but then the relative paths are:
'compute/ops/stats/orders/v2/', 'compute/ops/orders/v2/', and
'compute/ops/bulk/orders/v2/'. The auth API endpoint begins with the
same url and the relative path is 'v0/auth/'. Setting the base url for
auth to 'https://api.planet.com/v0/auth/' makes sense for the auth
client, but for the orders client we cannot use one of the relative
paths for the base url. We could use `compute/ops` but that doesn't
actually get us to the orders API endpoint. Backing off to just have the
base url be 'https://api.planet.com/' seems less confusing than having
the base url for the orders API be some portion of the full path to the
endpoint and having the base url for the auth API be the full path.