relax pin on httpx and respx, switch to anyio from pytest-asyncio, fix all warnings
To fix the 3.10 warnings we relax the pins on httpx and respx and tried
to relax the pin on pytest-asyncio but ultimately moved to anyio.

We switch to anyio from pytest-asyncio because the latest version of
pytest-asyncio kept throwing errors and anyio is already a dependency of httpx.

Warnings popped up due to not using a context manager for open(),
typically in json.load(open(filename)), so these were changed.

Warnings also popped up due to non-async cli tests using
pytest.mark.anyio, so these were removed.